Update license to include information about commercial use
In this commit, the license has been updated to include information about obtaining a commercial license for for-profit companies and individuals intending to use the software for commercial purposes. The commercial license allows source code modifications and permits proprietary use without the copyleft provisions of the GNU Affero General Public License version 3 (AGPLv3). The updated license also clarifies the distinction between non-commercial and commercial use of the software, providing users with the appropriate licensing options based on their intended use.